Background and Player elements added.
There are 3 background layers that move at different speeds.
They are on an endless loop. The BackgroundController needs more
work but we may have to cut the changing biomes idea. The player
jumps but I am working on a double jump ability which I think is
absolutely necessary. Im just having troubles with the raycasting.
I'm going to do more research tonight and hopefully have everything
ready by noon.

What we need to focus on tomorrow:
- The Obstacle Spawner (spawning obstacles at various heights)
- The different obstacles (spikes, blocks, blocks that drop)
- Animations
- SFX
- Music